ToDo creation will now go to the correct tab
When the ToDo form is up, the user can freely switch between
projects/inbox tabs and create a ToDo to that tab.

I was not going to make this a feature but seeing how the ToDo form
sorta needs a lot of information I felt that having the form NOT close
while switching tabs would save the user's time and frustration from
having to reenter the same information that was deleted by the tab
switch.